requirements.txt
Flask
gunicorn
mysql-connector-python
pymysql
psycopg2-binary